+++ app/controllers/bookingsync_portal/admin/base_controller.rb
@@ -3,6 +3,7 @@ require 'bookingsync_application/admin/common_base_controller'
 module BookingsyncPortal
   module Admin
     class BaseController < ApplicationController
+      helper BookingsyncPortal::Admin::ApplicationHelper
       layout 'bookingsync_portal/admin'
       respond_to :html
       include BookingsyncApplication::Admin::CommonBaseController