+++ spec/factories/connections.rb
@@ -1,9 +1,6 @@
 FactoryGirl.define do
   factory :connection, class: BookingsyncPortal.connection_model do
-    after(:build) do |connection, _evaluator|
-      account = create(:account)
-      connection.rental ||= build(:rental, account: account)
-      connection.remote_rental ||= build(:remote_rental, account: account)
-    end
+    rental
+    remote_rental { build(:remote_rental, account: @instance.rental.account) }
   end
 end